:bug: Fix bug in adjusting the cleanupPercentage control spinner (#1589)

+++ composeApp/src/commonMain/kotlin/com/crosspaste/ui/settings/StoreSettingsView.kt
@@ -596,7 +596,7 @@ fun StoreSettingsView() {
                 Counter(defaultValue = cleanupPercentage.toLong(), unit = "%", rule = {
                     it in 10..50
                 }) { currentCleanupPercentage ->
-                    configManager.updateConfig("cleanupPercentage", currentCleanupPercentage)
+                    configManager.updateConfig("cleanupPercentage", currentCleanupPercentage.toInt())
                 }
             }
         }
